Refactor: decompose BlueprintForm into sibling files
BlueprintForm.tsx was 1774 LOC — the largest file in the repo. Two
roughly-independent editor panels (conditional filter and JSON
schema) plus a pile of pure helpers made it expensive to touch. Pure
extraction — no behavior change. AJV validation stays in the parent
per the follow-ups doc.

- blueprint-schema-utils.ts: ajv instance plus the pure helpers
  (RELATIONSHIP_MAP, getRelationshipTypes, toTitleCase,
  PROPERTY_TYPES, STRING_FORMATS, toSlug, generateId,
  buildJsonSchema, schemaToProperties) that the schema and filter
  editors both read.
- BlueprintFilterEditor.tsx: the Conditional Filter block. Takes
  filter state/setters plus the already-fetched project-type and
  environment data as props so React Query behavior stays exactly
  the same (data fetches once in the parent).
- BlueprintSchemaEditor.tsx: the JSON Schema section — Visual/Code
  toggle, error banner, visual-mode property rows, advanced panel,
  UI maps, code-mode textarea. `enumRawText` and `toggleExpandProp`
  move into the child because they only touch local UI state; the
  four property mutation ops (addProperty, updateProperty,
  moveProperty, removeProperty) stay in the parent and pass down
  as props, because each calls syncVisualToCode which feeds
  rawSchema into AJV validation.

No BlueprintPreview component was extracted because no dedicated
preview panel exists in the current form — the Code mode is just a
textarea of the raw schema. If a preview panel lands later, it can
consume blueprint-schema-utils' buildJsonSchema.

Parent now 850 LOC. BlueprintSchemaEditor at 713 LOC is still large
and the advanced-options row + UI-map editors are good candidates
for further follow-up extraction (SchemaPropertyRow, UiMapEditor);
out of scope here.
